pyproject.toml
[project]
name = "telelgram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.61",
    "pandas>=2.2.3",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
]
